Don't merge &LABELS and CL:LABELS
Consider this example:

  (let+ (((&labels foo (...) ...)))
    (labels ((bar (...) ...))
      (do-something) ...)
    (bar ...))

The outer call to BAR is outside the scope of inner LABELS form, therefore
it is a call to a outer (possibly global) function named BAR. Previously,
that form would macroexpand into

  (labels ((foo (...) ...)
           (bar (...) ...))
    (do-something) ...
    (bar))

which makes (BAR) a call to the local function.

This commit is based on code by Jéssica Milaré with modifications by
Jan Moringen.